Add logout menu item
