+++ collectivo/auth/apps.py
@@ -1,5 +1,24 @@
 """Configuration file for the authentication module."""
 from django.apps import AppConfig
+from django.db.models.signals import post_migrate
+
+
+def post_migrate_callback(sender, **kwargs):
+    """Initialize extension after database is ready."""
+    from collectivo.extensions.utils import register_extension
+    from collectivo.ux.utils import register_menuitem
+
+    name = 'auth'
+    description = 'API for user authentication.'
+    register_extension(name=name, built_in=True, description=description)
+    register_menuitem(
+        item_id='auth_logout_button',
+        menu_id='main_menu',
+        label='Log out',
+        extension=name,
+        action='component',
+        component_name='logout'
+    )
 
 
 class AuthConfig(AppConfig):
@@ -8,3 +27,12 @@ class AuthConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'collectivo.auth'
     label = 'collectivo_auth'
+
+    def ready(self):
+        """
+        Initialize app when it is ready.
+
+        Database calls are performed after migrations, using the post_migrate
+        signal. This signal only works if the app has a models.py module.
+        """
+        post_migrate.connect(post_migrate_callback, sender=self)
