+++ collectivo/auth/models.py
@@ -1,0 +1,1 @@
+"""Models of the authentication module."""
