+++ collectivo/auth/tests/test_registration.py
@@ -1,15 +1,15 @@
-"""Tests of the members extension."""
+"""Tests of the authentication extension."""
 from django.test import TestCase
 from collectivo.extensions.models import Extension
 from collectivo.ux.models import MenuItem
 
 
-class MembersRegistrationTests(TestCase):
+class AuthRegistrationTests(TestCase):
     """Test that the extension is installed correctly."""
 
     def setUp(self):
         """Initialize testing instance."""
-        self.name = 'members'
+        self.name = 'auth'
 
     def test_extension_exists(self):
         """Test that the extension is automatically registered."""
@@ -19,8 +19,4 @@ class MembersRegistrationTests(TestCase):
     def test_menu_items_exist(self):
         """Test that the menu items are registered."""
         res = MenuItem.objects.filter(extension=self.name)
-        self.assertEqual(len(res), 2)
-
-    def test_menu_items_access(self):
-        """Test that only the correct roles can view menu item."""
-        # TODO Access control for menus
+        self.assertEqual(len(res), 1)
